添加查看详情

+++ ccf-system/src/main/java/com/xyz/modules/biz/rest/RentalhouseController.java
@@ -1,12 +1,14 @@
 package com.xyz.modules.biz.rest;
 
 import com.xyz.aop.log.Log;
+import com.xyz.exception.BadRequestException;
 import com.xyz.modules.biz.domain.Rentalhouse;
 import com.xyz.modules.biz.service.RentalhouseService;
 import com.xyz.modules.biz.service.dto.RentalhouseQueryCriteria;
 import com.xyz.modules.security.security.JwtUser;
 import com.xyz.modules.system.service.DeptService;
 import com.xyz.utils.SecurityUtils;
+import com.xyz.utils.StringUtils;
 import org.hibernate.annotations.DynamicUpdate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -53,13 +55,22 @@ public class RentalhouseController {
         criteria.setUnitCode(deptCodes);
         return new ResponseEntity(RentalhouseService.queryAll(criteria,pageable),HttpStatus.OK);
     }
+    @Log("查询单个Rentalhouse")
+    @ApiOperation(value = "查询单个Rentalhouse")
+    @GetMapping(value = "/Rentalhouse/{id}")
+    @PreAuthorize("hasAnyRole('ADMIN','RENTALHOUSE_ALL','RENTALHOUSE_SELECT')")
+    public ResponseEntity getById(@PathVariable String id){
+        if (StringUtils.isBlank(id)){
+            throw new BadRequestException("主键ID不能为空");
+        }
+        return new ResponseEntity(RentalhouseService.findById(id),HttpStatus.OK);
+    }
 
     @Log("新增Rentalhouse")
     @ApiOperation(value = "新增Rentalhouse")
     @PostMapping(value = "/Rentalhouse")
     @PreAuthorize("hasAnyRole('ADMIN','RENTALHOUSE_ALL','RENTALHOUSE_CREATE')")
     public ResponseEntity create(@Validated @RequestBody Rentalhouse resources){
-
         return new ResponseEntity(RentalhouseService.create(resources),HttpStatus.CREATED);
     }
 
@@ -68,6 +79,9 @@ public class RentalhouseController {
     @PutMapping(value = "/Rentalhouse")
     @PreAuthorize("hasAnyRole('ADMIN','RENTALHOUSE_ALL','RENTALHOUSE_EDIT')")
     public ResponseEntity update(@Validated @RequestBody Rentalhouse resources){
+        if (StringUtils.isBlank(resources.getRentId())){
+            throw new BadRequestException("主键ID不能为空");
+        }
         RentalhouseService.update(resources);
         return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
@@ -77,6 +91,9 @@ public class RentalhouseController {
     @DeleteMapping(value = "/Rentalhouse/{rentId}")
     @PreAuthorize("hasAnyRole('ADMIN','RENTALHOUSE_ALL','RENTALHOUSE_DELETE')")
     public ResponseEntity delete(@PathVariable String rentId){
+        if (StringUtils.isBlank(rentId)){
+            throw new BadRequestException("主键ID不能为空");
+        }
         RentalhouseService.delete(rentId);
         return new ResponseEntity(HttpStatus.OK);
     }
